Add RSpec and BDD version of unit tests
Gemfile now contains RSpec, so it is included with bundle install.
The test_report_generator tests and test_helper are left untouched.
report_generator_spec and file_util do the exact same thing, but
with a different syntax. Note the comments in report_generator_spec,
they should not persist. Also the readme should be updated.